Agrego la funcion de cambiar el tamaño del LeftPane y RightPane

+++ src/node/Separator.java
@@ -4,7 +4,7 @@ import javafx.geometry.Orientation;
 import javafx.scene.layout.Region;
 
 public class Separator extends Region {
-    public Separator(int size, Orientation orientation) {
+    public Separator(double size, Orientation orientation) {
         if (orientation.equals(Orientation.VERTICAL)) setPrefWidth(size);
         else setPrefHeight(size);
     }
